+++ tests/Propulsion.Kafka.Integration/ConsumersIntegration.fs
@@ -91,7 +91,7 @@ module Helpers =
 
             timeout |> Option.iter consumer.StopAfter
 
-            do! consumer.AwaitShutdown()
+            do! consumer.Await()
         }
 
         do! Async.Parallel [for i in 1 .. numConsumers -> mkConsumer i] |> Async.Ignore
@@ -142,7 +142,7 @@ module Helpers =
 
             timeout |> Option.defaultValue (TimeSpan.FromMinutes 15.) |> consumer.StopAfter
 
-            do! consumer.AwaitShutdown()
+            do! consumer.Await()
         }
 
         do! Async.Parallel [for i in 1 .. numConsumers -> mkConsumer i] |> Async.Ignore
@@ -180,7 +180,7 @@ module Helpers =
 
             timeout |> Option.defaultValue (TimeSpan.FromMinutes 15.) |> consumer.StopAfter
 
-            do! consumer.AwaitShutdown()
+            do! consumer.Await()
         }
 
         do! Async.Parallel [for i in 1 .. numConsumers -> mkConsumer i] |> Async.Ignore